requirements.txt
evernote3
oauth2
exif
Pillow
piexif